chore: Re-org so addLiquidityNested can switch on protocol version.
