+++ src/entities/addLiquidityNested/addLiquidityNestedV2/getQueryCallsAttributes.ts
@@ -1,17 +1,17 @@
-import { Token } from '../token';
+import { Address, PoolType } from '@/types';
+import { Token } from '@/entities/token';
 import {
-    BALANCER_RELAYER,
-    ChainId,
     ZERO_ADDRESS,
+    BALANCER_RELAYER,
     getPoolAddress,
-} from '../../utils';
+    ChainId,
+} from '@/utils';
 import {
-    AddLiquidityNestedInput,
     AddLiquidityNestedCallAttributes,
+    AddLiquidityNestedInput,
 } from './types';
-import { NestedPool, PoolKind } from '../types';
-import { Address, PoolType } from '../../types';
-import { Relayer } from '../relayer';
+import { NestedPool, PoolKind } from '@/entities/types';
+import { Relayer } from '@/entities/relayer';
 
 export const getQueryCallsAttributes = (
     { amountsIn, chainId, fromInternalBalance }: AddLiquidityNestedInput,
