+++ src/entities/priceImpact/index.ts
@@ -7,8 +7,6 @@ import {
     AddLiquiditySingleTokenInput,
     AddLiquidityUnbalancedInput,
 } from '../addLiquidity/types';
-import { AddLiquidityNested } from '../addLiquidityNested';
-import { AddLiquidityNestedInput } from '../addLiquidityNested/types';
 import { PriceImpactAmount } from '../priceImpactAmount';
 import { RemoveLiquidity } from '../removeLiquidity';
 import {
@@ -24,6 +22,8 @@ import { Token } from '../token';
 import { TokenAmount } from '../tokenAmount';
 import { NestedPoolState, PoolState } from '../types';
 import { getSortedTokens } from '../utils';
+import { AddLiquidityNestedInput } from '../addLiquidityNested/addLiquidityNestedV2/types';
+import { AddLiquidityNested } from '../addLiquidityNested';
 
 export class PriceImpact {
     /**
